requirements.txt
flask
tensorflow
numpy
opencv-python
Pillow
